feat: move submit POST call to platform (#3116)
