+++ src/app/shared/components/create-new-report-v2/create-new-report.component.spec.ts
@@ -35,12 +35,7 @@ describe('CreateNewReportComponent', () => {
 
   beforeEach(waitForAsync(() => {
     modalController = jasmine.createSpyObj('ModalController', ['dismiss']);
-    reportService = jasmine.createSpyObj('ReportService', [
-      'getReportPurpose',
-      'createDraft',
-      'addTransactions',
-      'create',
-    ]);
+    reportService = jasmine.createSpyObj('ReportService', ['getReportPurpose', 'createDraft', 'addTransactions']);
     trackingService = jasmine.createSpyObj('TrackingService', ['createReport']);
     refinerService = jasmine.createSpyObj('RefinerService', ['startSurvey']);
     currencyService = jasmine.createSpyObj('CurrencyService', ['getHomeCurrency']);
@@ -49,6 +44,7 @@ describe('CreateNewReportComponent', () => {
       'addExpenses',
       'createDraft',
       'suggestPurpose',
+      'create',
     ]);
     const humanizeCurrencyPipeSpy = jasmine.createSpyObj('HumanizeCurrency', ['transform']);
     const fyCurrencyPipeSpy = jasmine.createSpyObj('FyCurrencyPipe', ['transform']);
@@ -263,13 +259,13 @@ describe('CreateNewReportComponent', () => {
 
       const txnIds = ['txDDLtRaflUW', 'tx5WDG9lxBDT'];
       const report = expectedReportsSinglePage[0];
-      reportService.create.and.returnValue(of(expectedReportsSinglePage[0]));
+      spenderReportsService.create.and.returnValue(of(expectedReportsSinglePage[0]));
       component.ctaClickedEvent('submit_report');
       fixture.detectChanges();
       tick(500);
       expect(component.submitReportLoader).toBeFalse();
       expect(component.showReportNameError).toBeFalse();
-      expect(reportService.create).toHaveBeenCalledOnceWith(reportPurpose, txnIds);
+      expect(spenderReportsService.create).toHaveBeenCalledOnceWith(reportPurpose, txnIds);
       expect(refinerService.startSurvey).toHaveBeenCalledOnceWith({ actionName: 'Submit Newly Created Report' });
       expect(component.submitReportLoader).toBeFalse();
       expect(modalController.dismiss).toHaveBeenCalledOnceWith({
