+++ src/app/shared/components/create-new-report-v2/create-new-report.component.ts
@@ -148,7 +148,7 @@ export class CreateNewReportComponent implements OnInit {
         });
     } else {
       this.submitReportLoader = true;
-      this.reportService
+      this.spenderReportsService
         .create(report, txnIds)
         .pipe(
           tap(() => {
